removing duplicated code
